Add Animation Component Fixes #18
