+++ src/components/sections/partials/HeroAnimation.tsx
@@ -1,43 +1,11 @@
 import React from "react";
-import { useRef, useEffect } from "react";
-import { Rive } from "rive-js";
+import Animation from "components/elements/Animation";
 
 export default function HeroAnimation(): JSX.Element {
-  const canvas = useRef<HTMLCanvasElement>(null);
-  const container = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    const animation = new Rive({
-      src: "./reading_bot.riv",
-      canvas: canvas.current,
-      autoplay: true,
-    });
-
-    resize();
-    window.addEventListener("resize", resize);
-    return () => {
-      animation.current?.stop();
-      window.removeEventListener("resize", resize);
-    };
-  }, []);
-
-  const resize = () => {
-    if (container.current && canvas.current) {
-      const { width, height } = container.current.getBoundingClientRect();
-      canvas.current.width = width;
-      canvas.current.height = height;
-    }
-  };
-
   return (
-    <div
-      ref={container}
-      className="OuterContainer"
+    <Animation
+      src="./reading_bot.riv"
       style={{ height: "calc(100vh - 286px)" }}
-    >
-      <div className="InnerContainer">
-        <canvas ref={canvas} />
-      </div>
-    </div>
+    />
   );
 }
