+++ tsconfig.json
@@ -17,5 +17,5 @@
     "noEmit": true,
     "jsx": "react-jsx"
   },
-  "include": [".d.ts", "src"]
+  "include": ["src/.d.ts", "src"]
 }
